pyproject.toml
[project]
name = "marionette"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Melissa Autumn", email = "melissa@thunderbird.net" }
]
dependencies = [
    "marionette-driver>=3.4.0",
    "marionette-harness>=5.0.0",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/marionette"]
